improvements

+++ src/main/java/hadoop/project/IntSumReducerPropertyType.java
@@ -1,24 +1,23 @@
 package hadoop.project;
 
 import java.io.IOException;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 import org.apache.hadoop.mapreduce.Reducer.Context;
 
 public class IntSumReducerPropertyType
-            extends Reducer<Text, IntWritable, Text, IntWritable> {
+            extends Reducer<Text, LongWritable, Text, LongWritable> {
         public IntSumReducerPropertyType(){}
-        private IntWritable result = new IntWritable();
-        private int count = 0;
-        public void reduce(Text key, Iterable<IntWritable> values, Context context)
+        private LongWritable result = new LongWritable();
+        private long count = 0;
+        public void reduce(Text key, Iterable<LongWritable> values, Context context)
           throws IOException, InterruptedException {
-            int sum = 0;
-            for (IntWritable val : values) {
+            long sum = 0;
+            for (LongWritable val : values) {
               sum += val.get();
               count++;
             }
-        //context.write(key, new IntWritable(sum));
         result.set(sum/count);
         context.write(new Text("Property type: '"+key.toString()+"' Avg price:"), result);
         }
